src/main/java/kitchenpos/domain/OrderTable.java
package kitchenpos.domain;

public class OrderTable {

    private static final int MIN_GUESTS_COUNTS = 0;

    private Long id;
    private Long tableGroupId;
    private int numberOfGuests;
    private boolean empty;

    public OrderTable(final Long tableGroupId, final int numberOfGuests, final boolean empty) {
        this(null, tableGroupId, numberOfGuests, empty);
    }

    public OrderTable(final Long id, final Long tableGroupId, final int numberOfGuests, final boolean empty) {
        this.id = id;
        this.tableGroupId = tableGroupId;
        this.numberOfGuests = numberOfGuests;
        this.empty = empty;
    }

    public void updateNumberOfGuests(final int numberOfGuests) {
        if (empty) {
            throw new IllegalArgumentException("빈 테이블에는 손님을 지정할 수 없습니다.");
        }
        if (numberOfGuests < MIN_GUESTS_COUNTS) {
            throw new IllegalArgumentException("손님 수는 0보다 작을 수 없습니다.");
        }
        this.numberOfGuests = numberOfGuests;
    }

    public void updateStatusEmpty(final boolean empty) {
        this.empty = empty;
    }

    public Long getId() {
        return id;
    }

    public void update(final Long tableGroupId, final boolean empty) {
        this.tableGroupId = tableGroupId;
        this.empty = empty;
    }

    public Long getTableGroupId() {
        return tableGroupId;
    }

    public int getNumberOfGuests() {
        return numberOfGuests;
    }

    public boolean isEmpty() {
        return empty;
    }


}
